stagingtest/requirements.txt
mysql-connector-python
requests
selenium
unittest-xml-reporting